Global refactoring

+++ CDescriptor.cpp
@@ -35,24 +35,24 @@ void CDescriptor::setInterestPoint( QPoint _interestPoint )
 void CDescriptor::normalize()
 {
     float l2 = 0;
-    for( int i = 0; i < m_descriptor.size(); i++ )
+    for( size_t i = 0; i < m_descriptor.size(); i++ )
     {
         l2 += m_descriptor[i].sumOfSquares();
     }
     l2 = sqrt(l2);
-    for( int i = 0; i < m_descriptor.size(); i++ )
+    for( size_t i = 0; i < m_descriptor.size(); i++ )
     {
         m_descriptor[i].normalize( l2 );
         m_descriptor[i].bound(0.2);
     }
 
     l2 = 0;
-    for( int i = 0; i < m_descriptor.size(); i++ )
+    for( size_t i = 0; i < m_descriptor.size(); i++ )
     {
         l2 += m_descriptor[i].sumOfSquares();
     }
     l2 = sqrt( l2 );
-    for( int i = 0; i < m_descriptor.size(); i++ )
+    for( size_t i = 0; i < m_descriptor.size(); i++ )
     {
         m_descriptor[i].normalize( l2 );
     }
